gene selection refactoring
