+++ dynamo/preprocessing/preprocessor_utils.py
@@ -13,7 +13,6 @@ import scipy.sparse
 from anndata import AnnData
 from scipy.sparse import csr_matrix, spmatrix
 from scipy.sparse.base import issparse
-from sklearn.svm import SVR
 from sklearn.utils import sparsefuncs
 
 from ..configuration import DKM, DynamoAdataKeyManager
@@ -31,7 +30,6 @@ from ..dynamo_logger import (
 )
 from ..tools.utils import update_dict
 from ..utils import copy_adata
-from .preprocess_monocle_utils import top_table
 from .utils import (
     Freeman_Tukey,
     add_noise_to_duplicates,
@@ -251,688 +249,6 @@ def seurat_get_mean_var(
     mean = np.nan_to_num(mean)
     var = np.nan_to_num(var)
     return mean, var
-
-
-def select_genes_by_dispersion_general(
-    adata: AnnData,
-    layer: str = DKM.X_LAYER,
-    nan_replace_val: Union[float, None] = None,
-    n_top_genes: int = 2000,
-    recipe: Literal["monocle", "svr", "seurat"] = "monocle",
-    seurat_min_disp: Union[float, None] = None,
-    seurat_max_disp: Union[float, None] = None,
-    seurat_min_mean: Union[float, None] = None,
-    seurat_max_mean: Union[float, None] = None,
-    monocle_kwargs: dict = {},
-    gene_names: Union[List[str], None] = None,
-    var_filter_key: str = "pass_basic_filter",
-    inplace: bool = False,
-) -> None:
-    """A general function for feature genes selection.
-
-    Preprocess adata and dispatch to different filtering methods, and eventually set keys in anndata to denote which
-    genes are wanted in downstream analysis.
-
-    Args:
-        adata: an AnnData object.
-        layer: the key of a sparse matrix in adata. Defaults to DKM.X_LAYER.
-        nan_replace_val: your choice of value to replace values in layer. Defaults to None.
-        n_top_genes: number of genes to select as highly variable genes. Defaults to 2000.
-        recipe: a recipe for selecting genes; must be one of "monocle", "svr", or "seurat". Defaults to "monocle".
-        seurat_min_disp: seurat dispersion min cutoff. Defaults to None.
-        seurat_max_disp: seurat dispersion max cutoff. Defaults to None.
-        seurat_min_mean: seurat mean min cutoff. Defaults to None.
-        seurat_max_mean: seurat mean max cutoff. Defaults to None.
-        monocle_kwargs: kwargs for `select_genes_monocle`. Defaults to {}.
-        gene_names: name of genes to be selected. Defaults to None.
-        var_filter_key: filter gene names based on the key defined in adata.var before gene selection. Defaults to
-            "pass_basic_filter".
-        inplace: when inplace is True, subset adata according to selected genes. Defaults to False.
-
-    Raises:
-        NotImplementedError: the recipe is invalid/unsupported.
-    """
-
-    main_info("filtering genes by dispersion...")
-    main_log_time()
-
-    pass_filter_genes = adata.var_names
-    if gene_names:
-        main_info("select genes on gene names from arguments <gene_names>")
-        pass_filter_genes = gene_names
-    elif var_filter_key:
-        main_info("select genes on var key: %s" % (var_filter_key))
-        pass_filter_genes = adata.var_names[adata.var[var_filter_key]]
-
-    if len(pass_filter_genes) != len(set(pass_filter_genes)):
-        main_warning("gene names are not unique, please check your preprocessing procedure.")
-    subset_adata = adata[:, pass_filter_genes]
-    if n_top_genes is None:
-        main_info("n_top_genes is None, reserve all genes and add filter gene information")
-        n_top_genes = adata.n_vars
-    layer_mat = DKM.select_layer_data(subset_adata, layer)
-    if nan_replace_val:
-        main_info("replacing nan values with: %s" % (nan_replace_val))
-        _mask = get_nan_or_inf_data_bool_mask(layer_mat)
-        layer_mat[_mask] = nan_replace_val
-
-    main_info("select genes by recipe: " + recipe)
-    if recipe == "svr":
-        mean, variance, highly_variable_mask, highly_variable_scores = select_genes_by_dispersion_svr(
-            subset_adata, layer_mat, n_top_genes
-        )
-    elif recipe == "seurat":
-        mean, variance, highly_variable_mask, highly_variable_scores = select_genes_by_seurat_recipe(
-            subset_adata,
-            layer_mat,
-            min_disp=seurat_min_disp,
-            max_disp=seurat_max_disp,
-            min_mean=seurat_min_mean,
-            max_mean=seurat_max_mean,
-            n_top_genes=n_top_genes,
-        )
-    elif recipe == "monocle":
-        # TODO refactor dynamo monocle selection genes part code and make it modular (same as the two functions above)
-        # the logics here for dynamo recipe is different from the above recipes
-        # Note we do not need to pass subset_adata here because monocle takes care of everything regarding dynamo
-        # convention
-        select_genes_monocle(adata, **monocle_kwargs)
-        adata.var[DKM.VAR_GENE_HIGHLY_VARIABLE_KEY] = adata.var[DKM.VAR_USE_FOR_PCA]
-        return
-    else:
-        raise NotImplementedError("Selected gene seletion recipe not supported.")
-
-    main_info_insert_adata_var(DKM.VAR_GENE_MEAN_KEY)
-    main_info_insert_adata_var(DKM.VAR_GENE_VAR_KEY)
-    main_info_insert_adata_var(DKM.VAR_GENE_HIGHLY_VARIABLE_KEY)
-    main_debug("type of variance:" + str(type(variance)))
-    main_debug("shape of variance:" + str(variance.shape))
-    adata.var[DKM.VAR_GENE_MEAN_KEY] = np.nan
-    adata.var[DKM.VAR_GENE_VAR_KEY] = np.nan
-    adata.var[DKM.VAR_GENE_HIGHLY_VARIABLE_KEY] = False
-    adata.var[DKM.VAR_USE_FOR_PCA] = False
-
-    adata.var[DKM.VAR_GENE_MEAN_KEY][pass_filter_genes] = mean.flatten()
-    adata.var[DKM.VAR_GENE_VAR_KEY][pass_filter_genes] = variance
-    adata.var[DKM.VAR_GENE_HIGHLY_VARIABLE_KEY][pass_filter_genes] = highly_variable_mask
-    adata.var[DKM.VAR_USE_FOR_PCA][pass_filter_genes] = highly_variable_mask
-
-    main_info("number of selected highly variable genes: " + str(adata.var[DKM.VAR_USE_FOR_PCA].sum()))
-    if recipe == "svr":
-        # SVR can give highly_variable_scores
-        main_info_insert_adata_var(DKM.VAR_GENE_HIGHLY_VARIABLE_SCORES)
-        adata.var[DKM.VAR_GENE_HIGHLY_VARIABLE_SCORES] = np.nan
-        adata.var[DKM.VAR_GENE_HIGHLY_VARIABLE_SCORES][pass_filter_genes] = highly_variable_scores.flatten()
-
-    if inplace:
-        main_info("inplace is True, subset adata according to selected genes.")
-        adata = adata[:, adata.var[DKM.VAR_USE_FOR_PCA]]
-    main_finish_progress("filter genes by dispersion")
-
-
-def select_genes_by_seurat_recipe(
-    adata: AnnData,
-    sparse_layer_mat: csr_matrix,
-    n_bins: int = 20,
-    log_mean_and_dispersion: bool = True,
-    min_disp: float = None,
-    max_disp: float = None,
-    min_mean: float = None,
-    max_mean: float = None,
-    n_top_genes: Union[int, None] = None,
-) -> Tuple[np.ndarray, np.ndarray, np.ndarray, None]:
-    """Apply seurat's gene selection recipe by cutoffs.
-
-    Args:
-        adata: an AnnData object
-        sparse_layer_mat: the sparse matrix used for gene selection.
-        n_bins: the number of bins for normalization. Defaults to 20.
-        log_mean_and_dispersion: whether log the gene expression values before calculating the dispersion values.
-            Defaults to True.
-        min_disp: seurat dispersion min cutoff. Defaults to None.
-        max_disp: seurat dispersion max cutoff. Defaults to None.
-        min_mean: seurat mean min cutoff. Defaults to None.
-        max_mean: seurat mean max cutoff. Defaults to None.
-        n_top_genes: number of top genes to be evaluated. If set to be None, genes are filtered by mean and dispersion
-            norm threshold. Defaults to None.
-
-    Returns:
-        A tuple (mean, variance, highly_variable_mask, highly_variable_scores), where mean is the mean of the provided
-        sparse matrix, variance is the variance of the provided sparse matrix, highly_variable_mask is a bool array
-        indicating whether an element (a gene) is highly variable in the matrix. highly_variable_scores is always none
-        since the scores are not applicable to Seurat recipe.
-    """
-
-    # default values from Seurat
-    if min_disp is None:
-        min_disp = 0.5
-    if max_disp is None:
-        max_disp = np.inf
-    if min_mean is None:
-        min_mean = 0.0125
-    if max_mean is None:
-        max_mean = 3
-
-    mean, variance, dispersion = calc_mean_var_dispersion_sparse(sparse_layer_mat)
-    sc_mean, sc_var = seurat_get_mean_var(sparse_layer_mat)
-    mean, variance = sc_mean, sc_var
-    dispersion = variance / mean
-
-    if log_mean_and_dispersion:
-        mean = np.log1p(mean)
-        dispersion[np.equal(dispersion, 0)] = np.nan
-        dispersion = np.log(dispersion)
-
-    temp_df = pd.DataFrame()
-    temp_df["mean"], temp_df["dispersion"] = mean, dispersion
-
-    temp_df["mean_bin"] = pd.cut(temp_df["mean"], bins=n_bins)
-    disp_grouped = temp_df.groupby("mean_bin")["dispersion"]
-    disp_mean_bin = disp_grouped.mean()
-    disp_std_bin = disp_grouped.std(ddof=1)
-
-    # handle nan std
-    one_gene_per_bin = disp_std_bin.isnull()
-
-    disp_std_bin[one_gene_per_bin] = disp_mean_bin[one_gene_per_bin].values
-    disp_mean_bin[one_gene_per_bin] = 0
-
-    # normalized dispersion
-    mean = disp_mean_bin[temp_df["mean_bin"].values].values
-    std = disp_std_bin[temp_df["mean_bin"].values].values
-    variance = std**2
-    temp_df["dispersion_norm"] = ((temp_df["dispersion"] - mean) / std).fillna(0)
-    dispersion_norm = temp_df["dispersion_norm"].values
-
-    highly_variable_mask = None
-    if n_top_genes is not None:
-        main_info("choose %d top genes" % (n_top_genes), indent_level=2)
-        threshold = temp_df["dispersion_norm"].nlargest(n_top_genes).values[-1]
-        highly_variable_mask = temp_df["dispersion_norm"].values >= threshold
-    else:
-        main_info("choose genes by mean and dispersion norm threshold", indent_level=2)
-        highly_variable_mask = np.logical_and.reduce(
-            (
-                mean > min_mean,
-                mean < max_mean,
-                dispersion_norm > min_disp,
-                dispersion_norm < max_disp,
-            )
-        )
-    return mean, variance, highly_variable_mask, None
-
-
-def select_genes_by_dispersion_svr(
-    adata: AnnData, layer_mat: Union[np.array, csr_matrix], n_top_genes: int
-) -> Tuple[np.ndarray, np.ndarray, np.ndarray, np.ndarray]:
-    """Filters adata's genes according to layer_mat, and set adata's preprocess keys for downstream analysis
-
-    Args:
-        adata: an AnnData object.
-        layer_mat: the matrix used for select genes with shape of #cells X #genes.
-        n_top_genes: the number of genes to use.
-
-    Returns:
-        A tuple (mean, variance, highly_variable_mask, highly_variable_scores), where mean is the mean of the provided
-        sparse matrix, variance is the variance of the provided sparse matrix, highly_variable_mask is a bool array
-        indicating whether an element (a gene) is highly variable in the matrix, and highly_variable_scores is an array
-        storing the dispersion score for each gene.
-    """
-
-    main_debug("type of layer_mat:" + str(type(layer_mat)))
-    if issparse(layer_mat):
-        main_info("layer_mat is sparse, dispatch to sparse calc function...")
-        mean, variance, dispersion = calc_mean_var_dispersion_sparse(layer_mat)
-    else:
-        main_info("layer_mat is np, dispatch to sparse calc function...")
-        mean, variance, dispersion = calc_mean_var_dispersion_ndarray(layer_mat)
-
-    highly_variable_mask, highly_variable_scores = get_highly_variable_mask_by_dispersion_svr(
-        mean, variance, n_top_genes
-    )
-    variance = np.array(variance).flatten()
-
-    return mean.flatten(), variance, highly_variable_mask, highly_variable_scores
-
-
-# Highly variable gene selection function:
-def get_highvar_genes_sparse(
-    expression: Union[
-        np.ndarray,
-        scipy.sparse.csr_matrix,
-        scipy.sparse.csc_matrix,
-        scipy.sparse.coo_matrix,
-    ],
-    expected_fano_threshold: Optional[float] = None,
-    numgenes: Optional[int] = None,
-    minimal_mean: float = 0.5,
-) -> Tuple[pd.DataFrame, Dict]:
-    """Find highly-variable genes in sparse single-cell data matrices.
-
-    Args:
-        expression: Gene expression matrix
-        expected_fano_threshold: Optionally can be used to set a manual dispersion threshold (for definition of
-            "highly-variable")
-        numgenes: Optionally can be used to find the n most variable genes
-        minimal_mean: Sets a threshold on the minimum mean expression to consider
-
-    Returns:
-        gene_counts_stats: Results dataframe containing pertinent information for each gene
-        gene_fano_parameters: Additional informative dictionary (w/ records of dispersion for each gene, threshold,
-        etc.)
-    """
-    gene_mean = np.array(expression.mean(axis=0)).astype(float).reshape(-1)
-    E2 = expression.copy()
-    E2.data **= 2
-    gene2_mean = np.array(E2.mean(axis=0)).reshape(-1)
-    gene_var = pd.Series(gene2_mean - (gene_mean**2))
-    del E2
-    gene_mean = pd.Series(gene_mean)
-    gene_fano = gene_var / gene_mean
-
-    # Find parameters for expected fano line -- this line can be non-linear...
-    top_genes = gene_mean.sort_values(ascending=False)[:20].index
-    A = (np.sqrt(gene_var) / gene_mean)[top_genes].min()
-
-    w_mean_low, w_mean_high = gene_mean.quantile([0.10, 0.90])
-    w_fano_low, w_fano_high = gene_fano.quantile([0.10, 0.90])
-    winsor_box = (
-        (gene_fano > w_fano_low) & (gene_fano < w_fano_high) & (gene_mean > w_mean_low) & (gene_mean < w_mean_high)
-    )
-    fano_median = gene_fano[winsor_box].median()
-    B = np.sqrt(fano_median)
-
-    gene_expected_fano = (A**2) * gene_mean + (B**2)
-    fano_ratio = gene_fano / gene_expected_fano
-
-    # Identify high var genes
-    if numgenes is not None:
-        highvargenes = fano_ratio.sort_values(ascending=False).index[:numgenes]
-        high_var_genes_ind = fano_ratio.index.isin(highvargenes)
-        T = None
-    else:
-        if not expected_fano_threshold:
-            T = 1.0 + gene_fano[winsor_box].std()
-        else:
-            T = expected_fano_threshold
-
-        high_var_genes_ind = (fano_ratio > T) & (gene_mean > minimal_mean)
-
-    gene_counts_stats = pd.DataFrame(
-        {
-            "mean": gene_mean,
-            "var": gene_var,
-            "fano": gene_fano,
-            "expected_fano": gene_expected_fano,
-            "high_var": high_var_genes_ind,
-            "fano_ratio": fano_ratio,
-        }
-    )
-    gene_fano_parameters = {
-        "A": A,
-        "B": B,
-        "T": T,
-        "minimal_mean": minimal_mean,
-    }
-    return (gene_counts_stats, gene_fano_parameters)
-
-
-def SVRs(
-    adata_ori: anndata.AnnData,
-    filter_bool: Union[np.ndarray, None] = None,
-    layers: str = "X",
-    relative_expr: bool = True,
-    total_szfactor: str = "total_Size_Factor",
-    min_expr_cells: int = 0,
-    min_expr_avg: int = 0,
-    max_expr_avg: int = np.inf,
-    svr_gamma: Union[float, None] = None,
-    winsorize: bool = False,
-    winsor_perc: Tuple[float, float] = (1, 99.5),
-    sort_inverse: bool = False,
-    use_all_genes_cells: bool = False,
-) -> anndata.AnnData:
-    """Support Vector Regression to identify highly variable genes.
-
-    This function is modified from https://github.com/velocyto-team/velocyto.py/blob/master/velocyto/analysis.py
-
-    Args:
-        adata_ori: an AnnData object
-        filter_bool: A boolean array from the user to select genes for downstream analysis. Defaults to None.
-        layers: The layer(s) to be used for calculating dispersion score via support vector regression (SVR). Defaults
-            to "X".
-        relative_expr: A logic flag to determine whether we need to divide gene expression values first by size factor
-            before run SVR. Defaults to True.
-        total_szfactor: The column name in the .obs attribute that corresponds to the size factor for the total mRNA.
-            Defaults to "total_Size_Factor".
-        min_expr_cells: minimum number of cells that express the gene for it to be considered in the fit. Defaults to 0.
-        min_expr_avg: The minimum average of genes across cells required for gene to be selected for SVR analyses.
-            Defaults to 0.
-        max_expr_avg: The maximum average of genes across cells required for gene to be selected for SVR analyses. Genes
-            with average gene expression larger than this value will be treated as house-keeping/outlier genes. Defaults
-            to np.inf.
-        svr_gamma: the gamma hyper-parameter of the SVR. Defaults to None.
-        winsorize: Wether to winsorize the data for the cv vs mean model. Defaults to False.
-        winsor_perc: the up and lower bound of the winsorization. Defaults to (1, 99.5).
-        sort_inverse: whether to sort genes from less noisy to more noisy (to use for size estimation not for feature
-            selection). Defaults to False.
-        use_all_genes_cells: A logic flag to determine whether all cells and genes should be used for the size factor
-            calculation. Defaults to False.
-
-    Returns:
-        An updated annData object with `log_m`, `log_cv`, `score` added to .obs columns and `SVR` added to uns attribute
-        as a new key.
-    """
-
-    layers = DKM.get_available_layer_keys(adata_ori, layers)
-
-    if use_all_genes_cells:
-        # let us ignore the `inplace` parameter in pandas.Categorical.remove_unused_categories  warning.
-        with warnings.catch_warnings():
-            warnings.simplefilter("ignore")
-            adata = adata_ori[:, filter_bool].copy() if filter_bool is not None else adata_ori
-    else:
-        cell_inds = adata_ori.obs.use_for_pca if "use_for_pca" in adata_ori.obs.columns else adata_ori.obs.index
-        filter_list = ["use_for_pca", "pass_basic_filter"]
-        filter_checker = [i in adata_ori.var.columns for i in filter_list]
-        which_filter = np.where(filter_checker)[0]
-
-        gene_inds = adata_ori.var[filter_list[which_filter[0]]] if len(which_filter) > 0 else adata_ori.var.index
-
-        # let us ignore the `inplace` parameter in pandas.Categorical.remove_unused_categories  warning.
-        with warnings.catch_warnings():
-            warnings.simplefilter("ignore")
-            adata = adata_ori[cell_inds, gene_inds].copy()
-        filter_bool = filter_bool[gene_inds]
-
-    for layer in layers:
-        if layer == "raw":
-            CM = adata.X.copy() if adata.raw is None else adata.raw
-            szfactors = (
-                adata.obs[layer + "_Size_Factor"].values[:, None]
-                if adata.raw.X is not None
-                else adata.obs["Size_Factor"].values[:, None]
-            )
-        elif layer == "X":
-            CM = adata.X.copy()
-            szfactors = adata.obs["Size_Factor"].values[:, None]
-        elif layer == "protein":
-            if "protein" in adata.obsm_keys():
-                CM = adata.obsm["protein"].copy()
-                szfactors = adata.obs[layer + "_Size_Factor"].values[:, None]
-            else:
-                continue
-        else:
-            CM = adata.layers[layer].copy()
-            szfactors = (
-                adata.obs[layer + "_Size_Factor"].values[:, None]
-                if layer + "_Size_Factor" in adata.obs.columns
-                else None
-            )
-
-        if total_szfactor is not None and total_szfactor in adata.obs.keys():
-            szfactors = adata.obs[total_szfactor].values[:, None] if total_szfactor in adata.obs.columns else None
-
-        if szfactors is not None and relative_expr:
-            if issparse(CM):
-                sparsefuncs.inplace_row_scale(CM, 1 / szfactors)
-            else:
-                CM /= szfactors
-
-        if winsorize:
-            if min_expr_cells <= ((100 - winsor_perc[1]) * CM.shape[0] * 0.01):
-                min_expr_cells = int(np.ceil((100 - winsor_perc[1]) * CM.shape[1] * 0.01)) + 2
-
-        detected_bool = np.array(
-            ((CM > 0).sum(0) >= min_expr_cells) & (CM.mean(0) <= max_expr_avg) & (CM.mean(0) >= min_expr_avg)
-        ).flatten()
-
-        valid_CM = CM[:, detected_bool]
-        if winsorize:
-            down, up = (
-                np.percentile(valid_CM.A, winsor_perc, 0)
-                if issparse(valid_CM)
-                else np.percentile(valid_CM, winsor_perc, 0)
-            )
-            Sfw = (
-                np.clip(valid_CM.A, down[None, :], up[None, :])
-                if issparse(valid_CM)
-                else np.percentile(valid_CM, winsor_perc, 0)
-            )
-            mu = Sfw.mean(0)
-            sigma = Sfw.std(0, ddof=1)
-        else:
-            mu = np.array(valid_CM.mean(0)).flatten()
-            sigma = (
-                np.array(
-                    np.sqrt(
-                        (valid_CM.multiply(valid_CM).mean(0).A1 - (mu) ** 2)
-                        # * (adata.n_obs)
-                        # / (adata.n_obs - 1)
-                    )
-                )
-                if issparse(valid_CM)
-                else valid_CM.std(0, ddof=1)
-            )
-
-        cv = sigma / mu
-        log_m = np.array(np.log2(mu)).flatten()
-        log_cv = np.array(np.log2(cv)).flatten()
-        log_m[mu == 0], log_cv[mu == 0] = 0, 0
-
-        if svr_gamma is None:
-            svr_gamma = 150.0 / len(mu)
-        # Fit the Support Vector Regression
-        clf = SVR(gamma=svr_gamma)
-        # clf.fit(log_m[:, None], log_cv)
-
-        (gene_counts_stats, gene_fano_parameters) = get_highvar_genes_sparse(valid_CM)
-        target = np.array(gene_counts_stats["fano"]).flatten()
-        ground = np.array(gene_counts_stats["mean"]).flatten()[:, None]
-        clf.fit(ground, target)
-
-        fitted_fun = clf.predict
-        # ff = fitted_fun(log_m[:, None])
-        # score = log_cv - ff
-        ff = fitted_fun(ground)
-        score = target - ff
-        if sort_inverse:
-            score = -score
-
-        prefix = "" if layer == "X" else layer + "_"
-        (adata.var[prefix + "log_m"], adata.var[prefix + "log_cv"], adata.var[prefix + "score"],) = (
-            np.nan,
-            np.nan,
-            -np.inf,
-        )
-        (
-            adata.var.loc[detected_bool, prefix + "log_m"],
-            adata.var.loc[detected_bool, prefix + "log_cv"],
-            adata.var.loc[detected_bool, prefix + "score"],
-        ) = (
-            np.array(ground).flatten(),
-            np.array(target).flatten(),
-            np.array(score).flatten(),
-        )
-
-        key = "velocyto_SVR" if layer == "raw" or layer == "X" else layer + "_velocyto_SVR"
-        adata_ori.uns[key] = {"SVR": fitted_fun}
-
-    adata_ori = merge_adata_attrs(adata_ori, adata, attr="var")
-
-    return adata_ori
-
-
-def select_genes_monocle(
-    adata: anndata.AnnData,
-    layer: str = "X",
-    total_szfactor: str = "total_Size_Factor",
-    keep_filtered: bool = True,
-    sort_by: Literal["SVR", "gini", "dispersion"] = "SVR",
-    n_top_genes: int = 2000,
-    SVRs_kwargs: dict = {},
-    only_bools: bool = False,
-    exprs_frac_for_gene_exclusion: float = 1,
-    genes_to_exclude: Union[List[str], None] = None,
-) -> Union[anndata.AnnData, np.ndarray]:
-    """Select genes based on monocle recipe.
-
-    This version is here for modularization of preprocessing, so that users may try combinations of different
-    preprocessing procesudres in Preprocessor.
-
-    Args:
-        adata: an AnnData object.
-        layer: The data from a particular layer (include X) used for feature selection. Defaults to "X".
-        total_szfactor: The column name in the .obs attribute that corresponds to the size factor for the total mRNA.
-            Defaults to "total_Size_Factor".
-        keep_filtered: Whether to keep genes that don't pass the filtering in the adata object. Defaults to True.
-        sort_by: the sorting methods, either SVR, dispersion or Gini index, to be used to select genes. Defaults to
-            "SVR".
-        n_top_genes: the number of top genes based on scoring method (specified by sort_by) will be selected as feature
-            genes. Defaults to 2000.
-        SVRs_kwargs: kwargs for `SVRs`. Defaults to {}.
-        only_bools: Only return a vector of bool values. Defaults to False.
-        exprs_frac_for_gene_exclusion: threshold of fractions for high fraction genes. Defaults to 1.
-        genes_to_exclude: genes that are excluded from evaluation. Defaults to None.
-
-    Returns:
-        The adata object with genes updated if `only_bools` is false. Otherwise, the bool array representing selected
-        genes.
-    """
-
-    # The following size factor calculation is now a prerequisite for monocle recipe preprocess in preprocessor.
-    adata = calc_sz_factor(
-        adata,
-        total_layers=adata.uns["pp"]["experiment_total_layers"],
-        scale_to=None,
-        splicing_total_layers=False,
-        X_total_layers=False,
-        layers=adata.uns["pp"]["experiment_layers"],
-        genes_use_for_norm=None,
-    )
-
-    filter_bool = (
-        adata.var["pass_basic_filter"]
-        if "pass_basic_filter" in adata.var.columns
-        else np.ones(adata.shape[1], dtype=bool)
-    )
-
-    if adata.shape[1] <= n_top_genes:
-        filter_bool = np.ones(adata.shape[1], dtype=bool)
-    else:
-        if sort_by == "dispersion":
-            table = top_table(adata, layer, mode="dispersion")
-            valid_table = table.query("dispersion_empirical > dispersion_fit")
-            valid_table = valid_table.loc[
-                set(adata.var.index[filter_bool]).intersection(valid_table.index),
-                :,
-            ]
-            gene_id = np.argsort(-valid_table.loc[:, "dispersion_empirical"])[:n_top_genes]
-            gene_id = valid_table.iloc[gene_id, :].index
-            filter_bool = adata.var.index.isin(gene_id)
-        elif sort_by == "gini":
-            table = top_table(adata, layer, mode="gini")
-            valid_table = table.loc[filter_bool, :]
-            gene_id = np.argsort(-valid_table.loc[:, "gini"])[:n_top_genes]
-            gene_id = valid_table.index[gene_id]
-            filter_bool = gene_id.isin(adata.var.index)
-        elif sort_by == "SVR":
-            SVRs_args = {
-                "min_expr_cells": 0,
-                "min_expr_avg": 0,
-                "max_expr_avg": np.inf,
-                "svr_gamma": None,
-                "winsorize": False,
-                "winsor_perc": (1, 99.5),
-                "sort_inverse": False,
-            }
-            SVRs_args = update_dict(SVRs_args, SVRs_kwargs)
-            adata = SVRs(
-                adata,
-                layers=[layer],
-                total_szfactor=total_szfactor,
-                filter_bool=filter_bool,
-                **SVRs_args,
-            )
-            filter_bool = get_svr_filter(adata, layer=layer, n_top_genes=n_top_genes, return_adata=False)
-        # elif sort_by == "fano":
-
-    # filter genes by gene expression fraction as well
-    adata.var["frac"], invalid_ids = compute_gene_exp_fraction(X=adata.X, threshold=exprs_frac_for_gene_exclusion)
-    genes_to_exclude = (
-        list(adata.var_names[invalid_ids])
-        if genes_to_exclude is None
-        else genes_to_exclude + list(adata.var_names[invalid_ids])
-    )
-    if genes_to_exclude is not None and len(genes_to_exclude) > 0:
-        adata_exclude_genes = adata.var.index.intersection(genes_to_exclude)
-        adata.var.loc[adata_exclude_genes, "use_for_pca"] = False
-
-    if keep_filtered:
-        adata.var["use_for_pca"] = filter_bool
-    else:
-        adata._inplace_subset_var(filter_bool)
-        adata.var["use_for_pca"] = True
-
-    return filter_bool if only_bools else adata
-
-
-def get_highly_variable_mask_by_dispersion_svr(
-    mean: np.ndarray,
-    var: np.ndarray,
-    n_top_genes: int,
-    svr_gamma: Union[float, None] = None,
-    return_scores: bool = True,
-) -> Union[Tuple[np.ndarray, np.ndarray], np.ndarray]:
-    """Returns the mask with shape same as mean and var.
-
-    The mask indicates whether each index is highly variable or not. Each index should represent a gene.
-
-    Args:
-        mean: mean of the genes.
-        var: variance of the genes.
-        n_top_genes: the number of top genes to be inspected.
-        svr_gamma: coefficient for support vector regression. Defaults to None.
-        return_scores: whether returen the dispersion scores. Defaults to True.
-
-    Returns:
-        A tuple (highly_variable_mask, scores) where highly_variable_mask is a bool array indicating whether an element
-        (a gene) is highly variable in the matrix and scores is an array recording variable score for each gene. scores
-        would only be returned when `return_scores` is True.
-    """
-
-    # normally, select svr_gamma based on #features
-    if svr_gamma is None:
-        svr_gamma = 150.0 / len(mean)
-    from sklearn.svm import SVR
-
-    mean_log = np.log2(mean)
-    cv_log = np.log2(np.sqrt(var) / mean)
-    classifier = SVR(gamma=svr_gamma)
-    # fit&preidction will complain about nan values if not take cared here
-    is_nan_indices = np.logical_or(np.isnan(mean_log), np.isnan(cv_log))
-    if np.sum(is_nan_indices) > 0:
-        main_warning(
-            (
-                "mean and cv_log contain NAN values. We exclude them in SVR training. Please use related gene filtering "
-                "methods to filter genes with zero means."
-            )
-        )
-
-    classifier.fit(mean_log[~is_nan_indices, np.newaxis], cv_log.reshape([-1, 1])[~is_nan_indices])
-    scores = np.repeat(np.nan, len(mean_log))
-    # TODO handle nan values during prediction here
-    scores[~is_nan_indices] = cv_log[~is_nan_indices] - classifier.predict(mean_log[~is_nan_indices, np.newaxis])
-    scores = scores.reshape([-1, 1])  # shape should be #genes x 1
-
-    # score threshold based on n top genes
-    n_top_genes = min(n_top_genes, len(mean))  # maybe not enough genes there
-    score_threshold = np.sort(-scores)[n_top_genes - 1]
-    highly_variable_mask = scores >= score_threshold
-    highly_variable_mask = np.array(highly_variable_mask).flatten()
-    if return_scores:
-        return highly_variable_mask, scores
-    return highly_variable_mask
 
 
 def log1p_adata_layer(adata: AnnData, layer: str = DKM.X_LAYER, copy: bool = False) -> AnnData:
